pyproject.toml
[project]
name = "resume-matcher"
version = "0.1.0"
description = ""
authors = [
    {name = "Nevivaghani",email = "nevivaghani793@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "langchain (>=0.3.23,<0.4.0)",
    "requests (>=2.32.3,<3.0.0)",
    "python-docx (>=1.1.2,<2.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "pymupdf (>=1.25.5,<2.0.0)",
    "langchain-community (>=0.3.21,<0.4.0)",
    "langchain-google-genai (>=2.1.2,<3.0.0)",
    "langchain-groq (>=0.3.2,<0.4.0)",
    "pymongo (>=4.11.3,<5.0.0)",
    "streamlit (>=1.44.1,<2.0.0)",
    "faiss-cpu (>=1.10.0,<2.0.0)",
    "sentence-transformers (>=2.2.2,<3.0.0)",
    "transformers (>=4.31.0,<4.39.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
